requirements.txt
django>=4.2
djangorestframework
mysqlclient
djangorestframework-simplejwt
python-dotenv
